+++ app/src/main/java/org/librefit/db/entity/Set.kt
@@ -63,7 +63,7 @@ import kotlin.random.Random
 @Serializable
 data class Set(
     @PrimaryKey(true) val id: Long = Random.Default.nextLong(),
-    val load: Float = 0f,
+    val load: Double = 0.0,
     val reps: Int = 0,
     val elapsedTime: Int = 0,
     val completed: Boolean = false,
