+++ app/src/main/java/org/librefit/ui/models/UiSet.kt
@@ -33,7 +33,7 @@ import kotlin.random.Random
 @Immutable
 data class UiSet(
     val id: Long = Random.Default.nextLong() + System.currentTimeMillis(),
-    val load: Float = 0f,
+    val load: Double = 0.0,
     val reps: Int = 0,
     val elapsedTime: Int = 0,
     val completed: Boolean = false,
